add react-router and new route
