+++ src/pages/New.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function New() {
+  return <div>This is new</div>;
+}
